proto(): get the proof of concept working
this just hard codes a bunch of values to make sure that this thing
can talk to intellij. we hard code the props for the `my-component` cmp
that is generated from `npm init stencil`. upon installing this commit's
output into a stencil component starter project, we can observe that
types are properly being augmented in the HTML space