chore(repo): add renovate to the repo (#4)
keep dependencies up to date with renovate. this is a naive
configuration file, since we can't _really_ test this thing to verify
that it exists exactly how we want it to without landing it. consider
this a 'first pass' of sorts with known-good defaults that we've been
using in StencilJS