feat(html): implement go to declaration for custom elms
update `ElementInfo` types to store information about their backing
implementation class - both the symbol (class name) and the file that
contains said symbol. this information is pulled from stencil's
intermediate representation when the html contributions are generated.

in order to know how to resolve the file containing the class name
(symbol), the IDE needs to have a valid path to it. the path is derived
from the stencil configuration's `rootDir` field, which represents the
root of a stencil project. this value is propagated from the output
target's entrypoint in order to have a single source of truth. in the
future, we may decide to wire up the entire config object, instead of a
single field (or a new object with a subset of configuration fields),
depending on what additional config fields are needed. for now, this
ought to be "good enough", and allows us to pass a _definitely_ set
field, rather than a bag of fields with are optional/undefined.

the `validate` function has been implemented on this output target. it
verifies that the `rootDir` field is a string on the provided `config`
object. stencil _should_ be setting this field for users, and it is not
expected they set this themselves. as a result, no additional
documentation is needed at this time. should this field be unset for
some reason, `validate` will throw an error, and defer to stencil to
create a `Diagnostic` entity and handle it appropriately