feat(config): allow out file to be configurable
allow end users to decide where the generated json file is output. to do
so, we introduce a new optional config object to the project, which is
accepted at compile time (via a stencil config file). the config file
has a single optional property, which allows users to specify the
location the generated file should end up.

the output location is slightly opinionated:
- if not set, default the name to `web-types.json`
- if there is no json extension, add `web-types.json` to the provided
  path
- paths should be reconcilable against the stencil root dir