chore(ci): add example project tests (#26)
add a test to the project to verify that we haven't checked in an update
to the output of the output target. we do so by installing the built
output target into the project, running the build command, and checking
if any files have changed. if so, we report an error and exit with a
non-zero status code