docs(vue): add a Vue example file
Web types are not only useful for HTML files in IntelliJ IDEs, but Vue
files as well. This commit adds an example Vue file that mirrors the
HTML example file (`example/src/example.vue`).

This new Vue file is _not_ wired up to any development server/project.
Folks wishing to "test it out" are asked to open the file in an IntelliJ
IDE, as that's where the functionality of this project can be seen.

Misc. updates/rewordings are also included in this commit that I came
across while updating the documentation