chore(tests): fix internal typings for `ComponentCompilerMeta`
add `hasModernPropertyDecls`, set to `false`. we only use this for
testing, and this is not exposed to the user. this property was added
in https://github.com/ionic-team/stencil/pull/6076.

note: this is "our" (my) fault this broke, as I'm using an internal type
not meant to be exposed to users.